Make status bar thinner by using GtkBox instead of GtkStatusBar
I'm not sure how others but I personally just HATE the huge status bar
of GTK 3. It's such a colossal waste of precious vertical space and not
making the height adjustable is a terrible design choice IMO.

Anyway, end of ranting, the GTK status bar is just a GtkBox and a label
(plus some message stack that Geany doesn't even use) so it's no problem
to replace the statusbar with these and set the height to a sane amount
of pixels.

To me, the result is MUCH better and from the themes I tried, it looks
the same like the original status bar (apart from the height of course).

There might be a problem with plugins that try to access the status bar
directly and don't use msgwin_status_add(). The only plugin that seems
to do that (correctly) is the "web helper" plugin but from the code
it appears it falls back to creating a fake status bar if it doesn't
locate the one from Geany so it won't crash.