Use C parser for GLSL files and CUDA parser for CUDA files
The GLSL portion of c.c doesn't seem to do anything - it's just a synonym
for a C parser which can be used directly instead. This way we can
switch to the new cxx parser for GLSL too and reduce one more language
to care about in c.c.

For reference, the GLSL "parser" was introduced in commit

6ce421f1411060b8576bdac982cca11229b8ef96

Since we need to keep original identifiers for parsers and need some
parser instead of the removed GLSL parser, we can replace it with
the CUDA parser that is part of the new cxx parser and we can use
this parser for parsing CUDA files instead of C++ (the CUDA parser
is basically just C++ parsers with different keywords).