basic: use the upstream ctags parser
In the original "ctags sync" pull request I didn't notice an improvement
that was present in the Geany version. I submitted this change upstream
in

https://github.com/universal-ctags/ctags/pull/3249

and in addition fixed #2461.

I re-checked the pull request "sync uctags parsers with big changes" and
basic should be the only parser where something substantial was missed.

Fixes #2461.