Add tm_parser_scope_separator_printable()
The added tm_parser_scope_separator_printable() function is a human
readable representation of scope separator. For most languages, this is
equal to the scope separator but for some (mostly markup languages),
some "strange" scope separators are used like \0x3 or \"\" and we need
something more readable when displayed to users. " > " seems to look
good.

This new function can be used to:
1. Format calltips (though real languages use human-readable scope sep)
2. Symbol tree filter and display.