Allow mapping ctags "roles" to arbitrary tag manager type
This patch allows us to map roles reported by ctags to different tag
manager types. For instance, for the Go parser both packages and
imports are reported by the kind letter 'p' and are distinguished only
by the "imported" role.

(Code doing this for Go is commented-out for now because we need some
more improvements of the symbol tree because all root categories are
occupied and there's no place where to put the imports now.)