Remove pref "Use project-based session files", now always enabled
As per #267 we agreed that this pref makes little sense. Especially
as we're considering to move the project session files to a completely
separate file (so neither geany.conf nor $project.geany).

Currently, if not checked, the project session files would be stored
in the (new) session.conf file, overwriting the non-project session.
This is what #267 is about.