Disconnect default session and project session files
Currently, when loading "projects" (i.e. named sessions), the file list
overwrites the default session. Therefore, when loading a project, the
previously opened files are lost.

With this the default session is maintained separately. Any time a project
is opened it is written to session.conf and can be restored later on.