Update HACKING with up-to-date information about ctags parser addition
Previously "tagmanager" was a synonym for ctags+tagmanager but these
are separate now and tagmanager is more or less par of core Geany now
and not a separate library so update/remove things related to tagmanager.

Anjuta and exuberant ctags parsers aren't compatible with current
universal-ctags parsers (or the current tag manager) so they can be
removed from the description too.

The rest is just updates related to the move of stuff from symbols.c
to tm_parsers.c.