Statically allocate VteConfig (PR #3156)
In preparation of always setting up the VTE stash group, the global
VteConfig pointer is replaced by a plain global variable that does
not need to be allocated.

VteConfig is small and not loading libvte is probably a niche use case
so this change is generally beneficial by simplifying things.