Setup VTE stash group in init_pref_groups() (PR #3156)
Like with all other stash groups, the VTE one is now set up in
init_pref_groups(), along with the terminal various pref group.

Since be739e2 ("session.conf split follow-up #3"), the send_cmd_prefix
pref was read from the configuration before setting up the stash group,
which caused the "Follow path of the current file" feature to crash
Geany. I.e. the fix is to set up the stash group even earlier.

In my optionion, it's also beneficial that the overall stash groups
do not depend on loading libvte or not. For example, previously the
terminal various pref group was only added when loading libvte was not
disabled. When it was disabled the end result was inconsistent with
the manual. Now the terminal group appears unconditionally.

Fixes be739e2 ("session.conf split follow-up #3")
Fixes #3151