Add (and default to) tree view of open documents in the sidebar
Currently you can chose between a flat document list and a two-level
tree that groups the files by their directory.

With the new tree view the latter mode is extended to a true tree
where directory rows only represent the direct parent directory of
open documents. Except that single-child subtrees are collapsed to a single
row. See the manual for glory details.

The new mode allows to quickly close entire subtrees and is visually
close to the file system layout. As with the two-level tree, the home
directory and project base directory (if any project is open) are
treated specially and usually get a distict row
- home directory appears as "~"
- project base directory path is replaced by the project name

The new mode is the default. Existing installations will use the new
mode because a new pref is introduced.

Closes #259
Based on work by RPG <scriptum@users.noreply.github.com>