Enable local tag generation for C/C++
Enable generating these tags both for local variables and function
parameters - those are more or less identical for what we will be using
local tags for so they can be mapped to the same type.

Local tags aren't interesting for tag files so filter them out when
generating these (but this also means that we cannot create unit tests
for them).