Sort tags so tags from source file's header are preferred to other files
Since header files are always included by the corresponding source files,
it makes sense to query tags from headers before other file's tags.
To find header files, this patch assumes that they have the same
base name as the source files but with one of the C/C++ header extensions.

The patch also changes the order of operations to first remove
invalid local tags and then perform sorting to speed-up the sort as
this way the array contains less values.