Rewrite member_at_method_scope() to handle more situations
When performing scope completion for

void A::foo() {
    bar.  // <--
}

we need to determine what 'bar' is. It could be a global variable, in which
case we should look for variable tags, it could however also be a member
of A in which case we should look for member tags.

To determine this, the previous code checked whether there's a tag
named 'bar' with the same scope as the method tag. One drawback of this
approach is that it doesn't take namespace manipulation functions
like 'using namespace' into account so for header

namespace X {
    class A {
        Baz bar;
    };
};

and source

using namespace X;

void A::foo() {
    bar.  // <--
}

it wouldn't find 'bar' because ctags reports foo() to have scope A
and bar to have scope X::A.

Another drawback of this approach is that it doesn't take inheritance
into account so it wouldn't find 'bar' when defined in a super-class,
such as

class B {
    Baz bar;
};

class A : B {
    void foo();
}

To avoid these problems, this patch rewrites member_at_method_scope()
(and renames it to member_accessible()) so it gets the class name from
the scope of the method in which we are (A in the above example) and
returns all members of A (including the super-classes members).
Afterwards, it checks if one of the members is really the member tag
we are interested in ('bar' in the above example).