Speed up non-scope autocompletion
The previous code went through all the tags, moved all tags with the
typed prefix into a temporary array, moved "nearest" tags to the top,
removed duplicates and truncated the resulting array to the provided
max_num size.

The problem with this approach is that there may be too many tags with
the same prefix. For GTK and GLIB and the 'g' prefix we get about
30000 tags which then have to be processed, sorted, etc.

Instead of placing all tags with the given prefix into an array, we can
construct this array gradually by
1. First adding local tags from the current function
2. Then (unless we have reached max_num entries) adds tags from the current
file
3. Then from header
4. Then from workspace tags
5. Then from global tags

While adding, the code checks all the time whether max_num was reached
and when reached, it terminates the additions. (Note the comment
in the code about max_num which is doubled to avoid less tags in the
result array after final tag deduplication).