Statusbar Template enhancement: width specifier.
Every placeholder in the Statusbar Template may also specify width, e. g.
`%3l`. Width is the minimum number of characters to be shown. If the value is
shorter than width, the result is left-padded with spaces, or right-padded in
case of negative width. The value is not truncated even if the result is
larger.