Use constructor method calltip only if available
For D and Python, we try to read the calltip from the constructor method
(e.g. "__init__" for Python) instead of from the class tag itself.
If we don't find a matching constructor method tag, now we continue
looking at the class tag itself. This is useful for global tags where
the constructor signature might be defined on the class tag itself.