ctags: Matlab: function name only, without args
A line like `function y = foo(a, b, c)` should yield a tag of `foo`,
not `foo(a, b, c)`.
That way, Ctrl-clicking `foo` somewhere in the code will take me there.
The function name is `foo` after all, not `foo(a, b c)`.

Also, fixed issue where a line like `function foo() % this = bug`
would yield a tag of `bug` instead of `foo` because the `=` in the
comment was not ignored.