meson: Fix conditions when function is absent in Windows
This fixes the following error with meson in Windows
meson.build:441:32: ERROR: The  operator of bool does not accept objects of type int (1)

This set the value as integer (1/0) only. Previously, the value was
either integer (1/0) or boolean (true/false). In case of 'false',
the comparison becomes `if false == 0` which meson does not allow.