Add a flag indicating whether to trust isFileScope from ctags
ctags treats unknown C/C++ file extensions as if they were source
files. Because of this, the isFileScope flag is incorrectly set
to TRUE for such files even if they are headers and the defined
tags don't have a file-only scope.

To fix this, we intorduce a flag called trust_file_scope which we
set to TRUE only when the source file name matches one of the
known C/C++ extensions.

For all other languages the flag is set to TRUE.