Fix keyword colorization on filetype change
When filetype is changed to a compatible filetype (such as C->C++ and
then possibly back to C), colorization is lost. It seems that with
Scintilla 5, when language-specific highlighting changes, SCI_SETKEYWORDS
has to be called again to take effect.

By setting keyword_hash to 0 we force Geany to call sci_set_keywords() in
document_highlight_tags() which fixes the problem.