Enable new Scintilla "Change History"
Makes Geany feel a litttle bit more modern.

Can be enabled/disabled via Prefs. Default is to show the
change history in the markers margin.

Note: Scintilla tracks only its own change history and is not connected
to any VCS, unlike the git-changebar plugin. Also, change history is
lost when the document is closed and re-opened.