Don't comment global content unless it is Conf
When using Tools->Config Files to open a user file that doesn't exist.
This fixes the first line of geany.css being wrongly uncommented
(and the rest unchanged) due to toggle comment being used and CSS
not having a line comment.

(Verbatim copy-paste of Nick's description of the previous version
of the patch, thanks :-)