Prefix generated scope field by the "scope:" prefix
By default, ctags generates scope information using

parent_kind:A::B::C

where A::B::C is the scope of the parent tag. This is a bit problematic
when parsing such fields because we have to compare the parent_kind
value against all valid kinds of the language and if it matches, it
means that what follows is the scope. This slows down parsing of such
tags and also requires that parent_kind is really a valid kind for
the given languages.

ctags can also add extra "scope:" prefix to this field when using the
sZ field specification option on the command line:

scope:parent_kind:A::B::C

This way, the scope field can be detected easily by comparing against
"scope:" which speeds up parsing. Also, the parsing code doesn't care
about the value behind "scope:" so "class" can be used also for languages
that don't have the "class" kind.