Fix startup files order when placing tabs next to the current one
When the `tab_order_beside` option is enabled, tabs are opened on
either side of the active tab, depending on `tab_order_ltr`, rather
than at either end.

However, when loading startup files we don't switch to the last opened
document right away, but postpone this to the very end.  This affects
the order tabs are loaded, as the active one is always the first one at
this stage.

Rather than over-complicating the loading order to compensate the
various options' effect, temporarily revert to an set of options that
is easy to handle, and restore it after loading is finished.  This also
simplifies the existing logic as a basic forward iteration is now
enough.

Fixes #3609.