meson: Use only a single type for a given configuration data item
When a value could either be defined or not, we used to call either
`set(varname, 1)` or `set(varname, false)` respectively.  This however
leads to the value type of `varname` to be conditional, making it
virtually unusable for testing its value.
Instead, just do not emit anything when we used to emit `false`.

The only difference is that it will not generate any entry for the
disabled values, but that is what Autotools do already anyway --
although Autotools emit a comment.

This fixes build if either `HAVE_FNMATCH` or `HAVE_REGCOMP` is `false`
(that is, if either `fmatch()` or `regcomp()` functions are not found
on the system).

Fixes #3618.