Fix fairly nasty implicit integer conversion
Subtracting two unsigned values result in an unsigned value, and here
to use a different sign on each side of a ternary operator.

src/tagmanager/tm_workspace.c: In function 'sort_found_tags':
src/tagmanager/tm_workspace.c:909:45: warning: operand of '?:' changes signedness from 'int' to 'gulong' {aka 'long unsigned int'} due to unsignedness of other operand [-Wsign-compare]
  909 |                 return info->sort_by_name ? g_strcmp0(t1->name, t2->name) : t2->line - t1->line;
      |                                             ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~

This is an especially valid concern as TMTag::line is an unsigned long,
and the function returns an int, leading to possible sign swapping with
extreme values.  This is fairly theoretical as it's unlikely actual
files have enough lines for this to actually matter, but still.