"Fix" passing const arguments to spawn functions
`char**` doesn't coerce to `const char**` or any variant thereof, so
APIs expecting a `const char *const *` (like execv-style functions)
usually just use `char**` as it's usually effectively *more* convenient
for a caller.

However, this means that if we want to pass const data, we need a cast
(that is safe, as we know the API contract).

If making the compiler treat string literals as const (using e.g. GCC's
-Wwrite-strings), building an argument list composed of literals still
need to be const, so make it so and add a cast to the API call.