LSP interface for document symbol queries
Since the whole code of the symbol tree is quite complex, the idea
of the interface between Geany and LSP plugins is to reuse as much
of the Geany symbol tree code as possible. This means using the
TMTag structs by LSP plugins to pass the symbol data to Geany.

The main interface consists of 3 functions:
- lsp_doc_symbols_available(GeanyDocument *doc) - checking whether
  the plugin supports symbol tree queries for the document of the
  given file type
- void lsp_doc_symbols_request(GeanyDocument *doc,
  LspSymbolRequestCallback callback, gpointer user_data) - function
  performing asynchronous request to the LSP server to obtain
  symbols for the current document. Once available, callback is
  executed.
- GPtrArray *lsp_doc_symbols_get_cached(GeanyDocument *doc) - this
  function returns (synchronously) the symbols obtained by the last
  call of lsp_doc_symbols_request(). The return value of this function
  is an array of TMTag structures which are filled by the plugin
  as specified below.

TMTag:
- name - the name of the symbol
- type - the SymbolKind code as specified by LSP specification
- line - the line of the symbol
- arglist - the "detail" field of LSP document symbol (typically
  some form of function signature)
- scope - scope obtained from serializing the children field of
  LSP document symbol with the scope separator valid for the given
  filetype, obtained by symbols_get_context_separator()

In addition, there's a auxiliary function lsp_get_symbols_icon_id()
that maps LSP symbol kind (passed inside TMTag type member) to
a Geany symbol icon (TM_ICON_...).