Replace utils_make_human_readable_str() with g_format_size()
This allows removing a needlessly complex and tricky function (I dare
you understanding what it actually does the first time around) in which
we even introduced subtle (but luckily invisible for now) bugs over
time as it was probably not understood properly, and more sensitive to
breakage than it ought to be.

It also actually improves results, as we unit for bytes, and proper
translation for the units.

Note that this switches to SI units, which is probably actually better.