Do not use GtkResponseType as response ID type
`GtkResponseType` is supposed to be extensible by user code, and as
such is not used as a type in GTK public API, but as `gint`, which
allows passing other values (possibly from other enumerations) without
risking type conversion issues or compiler warnings.

Do the same in our own helper not to trigger GCC's `-Wenum-conversion`.